tidy: factor slack responses to new package
